+++ profiles/views.py
@@ -4,7 +4,7 @@ from django.shortcuts import render, get_object_or_404
 from django.contrib.auth.decorators import login_required
 
 from .models import UserProfile
-from .forms import UserProfileForm
+from .profile_form import UserProfileForm
 
 
 # pylint: disable=redefined-outer-name
