+++ requirements.txt
@@ -1,6 +1,7 @@
 asgiref==3.5.0
 boto3==1.21.32
 botocore==1.24.32
+coverage==6.3.2
 dj-database-url==0.5.0
 Django==3.2
 django-allauth==0.41.0
